Use webext-bridge for messaging
