+++ .gitignore
@@ -1,4 +1,4 @@
-manifest.json
 .idea/
-dist
 node_modules
+dist
+.webextrc
