--- bk/manifest-v3.json
@@ -1,46 +1,0 @@
-{
-  "manifest_version": 3,
-  "name": "Loadster Recorder",
-  "description": "Create test scripts to run in Loadster and Speedway.",
-  "version": "25",
-  "icons": {
-    "16": "images/icon-16x16.png",
-    "32": "images/icon-32x32.png",
-    "48": "images/icon-48x48.png",
-    "128": "images/icon-128x128.png",
-    "256": "images/icon-256x256.png"
-  },
-  "action": {
-    "default_icon": "images/icon-32x32.png",
-    "default_popup": "popup.html"
-  },
-  "background": {
-    "service_worker": "js/service-worker.js"
-  },
-  "minimum_chrome_version": "92",
-  "content_scripts": [
-    {
-      "matches": ["*://localhost/*", "https://loadster.app/*", "https://speedway.app/*"],
-      "js": [
-        "js/browser-polyfill.min.js",
-        "js/content.js"
-      ],
-      "run_at": "document_start"
-    }
-  ],
-  "web_accessible_resources": [
-    {
-      "resources": ["js/event-interceptor.js"],
-      "matches": ["<all_urls>"]
-    }
-  ],
-  "permissions": [
-    "storage",
-    "scripting",
-    "webRequest",
-    "webNavigation"
-  ],
-  "host_permissions": [
-    "*://*/*"
-  ]
-}
