+++ index.html
@@ -127,7 +127,7 @@
       }
 
     </style>
-<!--    <script src="js/popup.js"></script>-->
+    <script src="src/popup.js"></script>
   </head>
   <body>
     <div id="status-panel" class="panel">
