+++ package.json
@@ -5,6 +5,7 @@
   "description": "Create test scripts to run in Loadster.",
   "scripts": {
     "dev": "vite dev",
+    "dev:firefox": "cross-env TARGET=firefox vite dev",
     "build": "vite build",
     "build-chrome": "cross-env TARGET=chrome vite build",
     "build-firefox": "cross-env TARGET=firefox vite build",
