+++ src/constants.js
@@ -1,0 +1,14 @@
+export const RECORDING_STATUS = 'loadster.recording';
+export const RECORDING_EVENTS = 'RecordingEvents';
+export const RECORDING_STOP = 'RecordingStop';
+export const NAVIGATE_URL = 'Url';
+export const PING = 'Ping';
+export const PONG = 'Pong';
+export const USER_ACTION = 'loadster_browser_event';
+export const OPTIONS = 'loadster_recording_options';
+
+export const RECORDER_NAMESPACE = 'loadster-recorder-namespace';
+export const RECORDER_TYPE = {
+  HTTP: 'loadster-http-recorder',
+  BROWSER: 'loadster-browser-recorder'
+};
