+++ src/contentTab.js
@@ -1,0 +1,6 @@
+import { allowWindowMessaging } from 'webext-bridge/content-script';
+import { RECORDER_NAMESPACE } from './constants.js';
+
+console.log('client-tab-content.js');
+
+allowWindowMessaging(RECORDER_NAMESPACE);
